requirements.txt
discord.py==2.3.2
aiohttp==3.8.5
aiosqlite==0.19.0
wavelink==2.6.4
easy-pil==0.3.0
num2words==0.5.12